chore(agw): Move script-only dependencies to scripts folders
