+++ lte/gateway/python/scripts/cpe_monitoring_cli.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 """
 Copyright 2020 The Magma Authors.
