+++ lte/gateway/python/scripts/lte_health_service.py
@@ -22,12 +22,12 @@ from lte.protos.mobilityd_pb2 import IPAddress
 from lte.protos.mobilityd_pb2_grpc import MobilityServiceStub
 from magma.common.service_registry import ServiceRegistry
 from magma.configuration.mconfig_managers import load_service_mconfig_as_json
-from magma.health.entities import (
+from orc8r.protos.common_pb2 import Void
+from scripts.lte_health_entities import (
     AGWHealthSummary,
     CoreDumps,
     RegistrationSuccessRate,
 )
-from orc8r.protos.common_pb2 import Void
 
 
 class AGWHealth:
