+++ orc8r/gateway/python/scripts/health_cli.py
@@ -16,8 +16,8 @@ import os
 import subprocess
 
 import fire as fire
-from magma.common.health.docker_health_service import DockerHealthChecker
-from magma.common.health.health_service import GenericHealthChecker
+from scripts.docker_health_service import DockerHealthChecker
+from scripts.orc8r_health_service import GenericHealthChecker
 from termcolor import colored
