+++ orc8r/gateway/python/scripts/traffic_cli.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python3
+
 """
 Copyright 2020 The Magma Authors.
